dashboard_viewer/requirements.txt
beautifulsoup4==4.8.2
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
Django==2.2.13
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-constance==2.7.0
django-markdownify>=0.8.1
django-model-utils==4.0.0
django-sass==1.0.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
libsass==0.19.4
martor>=1.5.0
numpy==1.18.1
pandas==0.24.2
Pillow==7.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
redis>=3.5.3
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.2
sqlparse==0.3.0
urllib3==1.25.8